pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghostbrain"
version = "0.1.0"
description = "Personal second-brain automation for Obsidian"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Jannik", email = "jannik@codeship.tech" },
]
dependencies = [
    "anthropic>=0.40.0",
    "watchdog>=4.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
ghostbrain-bootstrap = "ghostbrain.bootstrap:main"
ghostbrain-worker = "ghostbrain.worker.main:main"
ghostbrain-claude-md = "ghostbrain.profile.claude_md:main"
ghostbrain-digest = "ghostbrain.worker.digest:main"
ghostbrain-github-fetch = "ghostbrain.connectors.github.__main__:main"
ghostbrain-jira-fetch = "ghostbrain.connectors.jira.__main__:main"
ghostbrain-confluence-fetch = "ghostbrain.connectors.confluence.__main__:main"
ghostbrain-profile-apply = "ghostbrain.profile.apply:main"
ghostbrain-profile-decay = "ghostbrain.profile.decay:main"

[tool.setuptools.packages.find]
include = ["ghostbrain*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
